Feat: Guard all necessary routes

+++ src/users/users.controller.ts
@@ -6,10 +6,14 @@ import {
   Param,
   Patch,
   Post,
+  Request,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersService } from './users.service';
+import { AuthGuard } from 'src/auth/auth.guard';
 
 @Controller('users')
 export class UsersController {
@@ -20,23 +24,39 @@ export class UsersController {
     return this.usersService.create(createUsersDto);
   }
 
+  @UseGuards(AuthGuard)
   @Get()
   findAll() {
     return this.usersService.findAll();
   }
 
+  @UseGuards(AuthGuard)
+  @Get('profile')
+  findMe(@Request() req: any) {
+    return this.usersService.findOne(+req.user.id);
+  }
+
+  @UseGuards(AuthGuard)
   @Get(':id')
   findOne(@Param('id') id: string) {
     return this.usersService.findOne(+id);
   }
 
+  @UseGuards(AuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUsersDto: UpdateUserDto) {
+  update(
+    @Param('id') id: string,
+    @Body() updateUsersDto: UpdateUserDto,
+    @Request() req: any,
+  ) {
+    if (+id !== +req.user.id) throw new UnauthorizedException();
     return this.usersService.update(+id, updateUsersDto);
   }
 
+  @UseGuards(AuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @Request() req: any) {
+    if (+id !== +req.user.id) throw new UnauthorizedException();
     return this.usersService.remove(+id);
   }
 }
